move unmanaged info to separate file
